Projects
Done in junior year with a team.  For their privacy, I have elected to
only leave their initials. All of them were coded in C. While I am
confident that Lab 4 is the final version, Labs 2 and 3 may not be.
Lab 2: I am very fuzzy what the exact purpose of this was for other than
to illustrate how parent and child processes work.
Lab 3: I believe this creates a heap and allows one to look at how
memory is allocated in it.
Lab 4: A basic proxy implementation to be used with a browser